+++ openfisca_france_data/erfs_fpr/scenario.py
@@ -1,11 +1,15 @@
 # -*- coding: utf-8 -*-
 
+from typing import Any
 
 from openfisca_france_data.surveys import AbstractErfsSurveyScenario
 
 
 class ErfsFprSurveyScenario(AbstractErfsSurveyScenario):
+    '''Survey scenario spécialisé pour l'ERFS-FPR.'''
+
     collection = 'openfisca_erfs_fpr'
+    # Les variables OpenFisca qu'on va utiliser avec les données en entrée.
     used_as_input_variables = [
         'activite',
         'autonomie_financiere',
@@ -34,11 +38,20 @@ class ErfsFprSurveyScenario(AbstractErfsSurveyScenario):
         'traitement_indiciaire_brut',
         'zone_apl',
         ]
+
     # Might be used: hsup
+    # Pour forcer quelques valeurs par défaut spécifiques.
     default_value_by_variable = dict(
         cotisation_sociale_mode_recouvrement = 2,
         # taux_incapacite = .50,
         )
+
+    # En général, si les données contiennent les variables dont on a besoin, on les garde.
+    #
+    # Le cas échéant, on les neutralise à leur valeur par défaut, pour ne pas garder des
+    # vecteurs des variables qu'on ne veut pas calculer.
+    #
+    # Ici c'est une exception : on déclare les variables qu'on ne veut pas neutraliser.
     non_neutralizable_variables = [
         'menage_ordinaire',
         'idfam_original',
@@ -49,12 +62,11 @@ class ErfsFprSurveyScenario(AbstractErfsSurveyScenario):
         'wprm_init',
         ]
 
-    def __init__(self, year = None):
-        assert year is not None
+    def __init__(self, year: int):
         self.year = year
 
     @classmethod
-    def build_input_data(cls, year = None):
-        assert year is not None
+    def build_input_data(cls, year: int) -> Any:
+        # TODO: fix import, otherwise it is untestable.
         from openfisca_france_data.erfs_fpr.input_data_builder import build
         build(year = year)